refactor(api): merge oidc_auth into auth, files+images into resources (#2849)
